changed int to double

+++ Addition.java
@@ -1,6 +1,6 @@
 public class Addition implements Operation{
 	public static String operation(double a, double b){
-		int sum=a+b;
+		double sum=a+b;
 		return "The addition is"+sum;
 	}
 }
